Improved some instructions
Shane set up an instance for me and I tested the commands up to line 20 and added some more explanation to some. Most of the explanations seemed fine, and the commands worked.

Some comments/questions (by NEW line number, with my added lines):

On line 26, you start listing the detailed steps. Can a developer skip the rest of the instructions if they don't need to customize things--does deploy.sh do all the steps for you? Or do they skip to line 83?

On line 30, please check that my rewrite is ok.

On line 43, it says "Start Base services". Is this supposed to say "KBase"? If not, why is Base capitalized (when services isn't)? Does "Base" mean something special here, or do you just mean start some of the services that are required by the KBase software (Mongo and mysql, here)?