Update preflight
Please do not merge yet!

I think that line 16 is trying to determine whether boot2docker is available, and if not, then run the curl to localhost.  With boot2docker being rolled in to Docker Toolbox, that environment variable no longer exists.  I want to check for DOCKER_HOST as well.